+++ lib/core/configuration.js
@@ -4,9 +4,6 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 exports.Configuration = void 0;
-var _stimulus = require("@hotwired/stimulus");
-var _form_controller = _interopRequireDefault(require("../controllers/form_controller"));
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }
 function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, "prototype", { writable: false }); return Constructor; }
@@ -25,10 +22,6 @@ var Configuration = /*#__PURE__*/function () {
           this[key] = value;
         });
       }
-      if (this.useStimulus) {
-        var application = _stimulus.Application.start();
-        application.register('hellotext--form', _form_controller.default);
-      }
       return this;
     }
   }, {
@@ -43,4 +36,3 @@ exports.Configuration = Configuration;
 Configuration.apiRoot = 'https://api.hellotext.com/v1';
 Configuration.autoGenerateSession = true;
 Configuration.autoMountForms = true;
-Configuration.useStimulus = true;
