+++ src/index.js
@@ -1,2 +1,13 @@
+import { Application } from '@hotwired/stimulus'
+import Hellotext from './hellotext'
+
+import FormController from './controllers/form_controller'
+
+const application = Application.start()
+application.register('hellotext--form', FormController)
+
 import '../styles/index.css'
-export default from './hellotext'
+
+window.Hellotext = Hellotext
+
+export default Hellotext
