+++ src/vanilla.js
@@ -1,0 +1,1 @@
+export default from './hellotext'
